server/requirements.txt
fastapi==0.100.0
numpy==1.23.3
pydantic==2.0.3
pymongo==4.4.1
python-dotenv==1.0.0
uvicorn==0.23.1
